Add Broyden's method for root finding
`math.optim.broyden(function, x0, inv_J0)`

Also move conjugate_gradient to phi.math.optim
